Refactored performance.py - broke services into individual files. This does not change the way external users would import and use the performance services however.
